Use 'go install' instead of 'go get' for installing binaries.
go get is deprecated for installing binaries

Also I have created a single target for installing all the binaries,
other targets won't install binaries themselves as this is not an
expected behavior from the user side.

Also removed the installation of
sigs.k8s.io/controller-runtime/tools/setup-envtest as it is not used
anyware.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>